Introduce event handlers and dispatching
The script can now register event handlers through `Game.on(eventName,
callbackFunction)`. The event names can be found in `EventNames.h`.
Cosmo will dispatch events through `Game::dispatch_event`, where you
pass it an event name, and the arguments to pass directly to all the
event handlers for that event name.

Current biggest problem with the implementation is that there is no
communication between the handler(s) and the dispatcher. This will be
needed for the handler(s) to make actions based on what they decide. The
best example is currently in the `PlayerConnect` event. This event is
cancelable (by returning `false` in `Plugin::on_player_connect`), and
can also be given a rejection message to send to the client. This should
be possible from event handlers.