Check if `this` is an object of type `Game` in Game::on
Previously, we assumed the `this` object was an object of type `Game`.
This would cause a crash if Game.on was, for example, called like so:
```js
Game.on.call(1)
```

This also moves retrieving the `this` value to the start of the
function, as the `this` value is the first argument to `call`.