Check if `this` is an object of type `Command` in Command::register_
Previously, we assumed the `this` object was an object of type
`Command`. This would cause a crash if Command.register was,
for example, called like so:
```js
Game.Command.register.call(1)
```

This also moves retrieving the `this` value to the start of the
function, as the `this` value is the first argument to `call`.