Introduce `get_value_by_type_description<T>` and `GetClassname`
We can now ask an entity to calculate the pointer to a value based on a
`typedescription_t` field, from a data map. It simply adds the offset
given and returns that pointer.

We now use this to get the address to `m_iClassname`, in order to
re-implement `CBaseEntity::GetClassname`, and expose this in script as
the `classname` getter.