Introduce `EntityPrototype.position` getter and setter
To do this, we need to re-implement `CBaseEntity::GetAbsOrigin`, using
it's own datamap to find the `m_vecAbsOrigin` pointer.

To expose `SourceVector`s (and `QAngle`s) in scripts, we introduce:
- `to_source_vector`    - `JS::Value` to `SourceVector`
- `to_qangle`           - `JS::Value` to `QAngle`
- `from_source_vector`  - `SourceVector` to `JS::Array`

However, for the setter, we need to do a bit more work. We usually care
more about an entity's absolute (world) origin, but we can't set that
directly with our pointer to it from the data descriptor, simply because
entities don't really use it -- they use their local origin (to their
parent). Instead, we let `CBaseEntity::Teleport` do all the work for us,
as a virtual method. `Teleport` can change the absoulte origin, angles,
and velocity of an entity.