Allow a prototype `JS::Object` to be specified for `Entity`
Instead of taking a `GlobalObject` and assuming our prototype is
`EntityPrototype`, allow a constructor to state what their prototype is
explicitly.

`Entity` was made to not be `final` anymore, so it can actually be
inherited.

The `Entity::entity` getter was made to be templated, to allow returning
of a derived `CBaseEntity` type. The default was made to be
`CBaseEntity`, to keep existing usage.