Add `TeamRoundTimerPrototype.paused`
Although we can pause and resume the timer with some virtual methods, we
can't actually query the paused-ness of it. We can utilize it's
`SendTable` to get an offset to `m_bTimerPaused`. We utilize
`get_value_by_send_property` (similar to
`get_value_by_type_description`), to get this value from an entity.