Introduce `Signature`, to perform signature-scanning within a binary
Signature scanning is a technique used to find a memory address, based
on it's "signature", aka. the bytes contained at that address. We use
'?' to specify "no value", in cases where that value can change between
compilations (ex: a global pointer, the offset to a call, etc). By using
enough bytes from a memory address, we can guarantee a unqiue address
lookup, and (more importantly) future proof for when the binary is
changed: the pointer will change, but the signature should be able to
find it again.