Introduce `EntityPrototype.emitSound` and `Game.Server.emitSound`
We find the `IEngineSound` interface to get access to `EmitSound`. The
least the script needs to give us is the path to the sound (example:
`misc/banana_slip.wav`), and a source. A source can either be an entity,
which the sound will follow, or a world position. The script can also
change the sound's volume and pitch. The script can also give us a
recipient filter, to know who should hear the sound.