Move `map` getter/setter and `sayText2` function from `Game` to `Server`
The placement of these didn't make much sense, so let's put them in
server instead.

`on` can stay in `Game`, as the server doesn't provide it, it's more of
a meta system.
`modDescription`, although provided by the server, also describes the
platform the script is running on, so it can conditionally execute based
on the mod running. (TF2, etc)