Introduce `{get,set}SendPropertyValue` and `{get,set}DataFieldValue`
Networkable entities have a `SendTable`, which is used in the
implementation of network variables in Source. These can be integers,
floats, vectors, strings, and more. Sometimes, these types are extended
further: to represent entities, handles are used, but are encoded as
integers. To get the `Entity` back from this handle, we introduce
`Game.Server.getEntityByHandle`.

Data descriptions/datamaps are used to provide meta information about a
type's fields. Although used mostly for entities, this is not required,
and is occasionally used for non-entities in the Engine.