Implement `CBaseHandle::Get`
This is defined in the Engine, so we have to implement our own.
Thankfully, we do it very similar to the Engine with what we have
available. Although an entity index may match different entities between
frames, it's serial number will not -- we check this to tell if we're
talking about the same entity.