Use `CBaseHandle` instead of `CBaseEntity*` in `Scripting::Entity`
Keeping a `CBaseEntity*` around for longer than a single frame was a bit
scary. At any point it could be destroyed, and a script trying to use it
would most likely end up crashing, as a UAF. Instead, we keep a
`CBaseHandle`, which keeps an entity index AND a serial number. Although
entity indexes are re-used, serial numbers for each entity index are not
reused. This let's us know which entity we're talking about, and if it
is still the same one as when we created it.

Anything requiring the `Entity` in `EntityPrototype` (and derivatives)
will now use `ensure_this_entity`. This will check if the handle's
entity is valid, and if not, throw.

Script can use `EntityPrototype.isValid` to check the if an entity is
still valid, without throwing.